+++ tests/security/http/test_security_http_bearer_description.py
@@ -26,14 +26,14 @@ def test_security_http_bearer():
 def test_security_http_bearer_no_credentials():
     with create_client(routes=[read_current_user]) as client:
         response = client.get("/users/me")
-        assert response.status_code == 403, response.text
+        assert response.status_code == 401, response.text
         assert response.json() == {"detail": "Not authenticated"}
 
 
 def test_security_http_bearer_incorrect_scheme_credentials():
     with create_client(routes=[read_current_user]) as client:
         response = client.get("/users/me", headers={"Authorization": "Basic notreally"})
-        assert response.status_code == 403, response.text
+        assert response.status_code == 401, response.text
         assert response.json() == {"detail": "Invalid authentication credentials"}