+++ docs/en/docs/release-notes.md
@@ -9,12 +9,13 @@ hide:
 
 ### Added
 
-- - [add_asgi_app()](./routing/router.md#add_asgi_app) is now supported and serves as an utility if you want to add any ASGI
+- [add_asgi_app()](./routing/router.md#add_asgi_app) is now supported and serves as an utility if you want to add any ASGI
 compatible application into a Lilya/ChildLilya instance.
 
 ### Changed
 
 - Remove noisy warning on loading directives.
+- Expose `original` property from the `@scheduler` wrapper.
 
 ## 0.3.3